pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "discord-py>=2.5.0",
    "flask>=3.1.0",
    "python-dotenv>=1.0.1",
    "trafilatura>=2.0.0",
    "twilio>=9.4.6",
]
